#39: Final changes to front end composition and filetypes
