+++ components/graphics/listbuddy-task-svg.tsx
@@ -1,0 +1,26 @@
+
+function ListBuddyTaskSVG() {
+    return (
+        <svg width="150" height="180" viewBox="0 0 621 789" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <rect x="88.216" y="136.289" width="443.753" height="217.011" rx="16.0393" stroke="white" strokeWidth="32.0785"/>
+            <rect x="16.0393" y="411.147" width="588.107" height="360.882" rx="16.0393" stroke="white" strokeWidth="32.0785"/>
+            <line x1="80.1963" y1="288.881" x2="537.315" y2="288.881" stroke="white" strokeWidth="32.0785"/>
+            <line x1="168.412" y1="279.152" x2="168.412" y2="369.34" stroke="white" strokeWidth="32.0785"/>
+            <line x1="518.604" y1="505.63" x2="181.779" y2="505.63" stroke="white" strokeWidth="32.0785"/>
+            <line x1="518.603" y1="589.375" x2="181.778" y2="589.375" stroke="white" strokeWidth="32.0785"/>
+            <line x1="518.604" y1="670.974" x2="181.779" y2="670.974" stroke="white" strokeWidth="32.0785"/>
+            <line x1="459.793" y1="279.152" x2="459.793" y2="369.34" stroke="white" strokeWidth="32.0785"/>
+            <line x1="358.211" y1="279.152" x2="358.211" y2="369.34" stroke="white" strokeWidth="32.0785"/>
+            <line x1="261.975" y1="279.152" x2="261.975" y2="369.34" stroke="white" strokeWidth="32.0785"/>
+            <line x1="310.093" y1="47.241" x2="310.093" y2="137.429" stroke="white" strokeWidth="32.0785"/>
+            <ellipse cx="169.749" cy="209.364" rx="41.4348" ry="33.2836" fill="white"/>
+            <ellipse cx="114.949" cy="502.474" rx="29.4053" ry="23.6206" fill="white"/>
+            <ellipse cx="114.949" cy="586.22" rx="29.4053" ry="23.6206" fill="white"/>
+            <ellipse cx="114.948" cy="667.818" rx="29.4053" ry="23.6206" fill="white"/>
+            <ellipse cx="447.763" cy="209.364" rx="41.4348" ry="33.2836" fill="white"/>
+            <ellipse cx="310.093" cy="33.2836" rx="42.7714" ry="33.2836" fill="white"/>
+        </svg>
+    );
+}
+
+export { ListBuddyTaskSVG };
