+++ components/views/TaskView.tsx
@@ -1,7 +1,14 @@
-// components/TaskView.js
 import { useState, useEffect } from 'react';
+import { ListBuddyTaskSVG } from '../graphics/listbuddy-task-svg';
 
-export default function TaskView({ task }) {
+// TODO: clean up comments
+
+/**
+ * Creates a floating view that displays the current task with text and a simplify button.
+ * @param task Task object
+ * @returns Rendered task view
+ */
+export default function TaskView({ task }: { task: any }) {
   const [simplifiedTask, setSimplifiedTask] = useState(null); // State to store simplified task
   // const [userId, setUserId] = useState(null); // State to store data from sessionStorage
 
@@ -49,13 +56,14 @@ export default function TaskView({ task }) {
   };
 
   return (
-    
       <div className= {`p-5 ${color[0]} text-white rounded-xl h-full relative`}>
           <h2 className="text-3xl font-bold">{task.name}</h2>
           <p className="mt-4 text-sm font-bold">Complete by: {new Date(task.deadline).toLocaleDateString()}</p>
           <p className="mt-4 font-bold">{simplifiedTask ? simplifiedTask : task.description}</p>
           <div className="flex items-center space-x-2 absolute bottom-4 justify-center mt-6">
           <button
+              // TODO: make this dynamic for if an API for AI is connected or not
+              disabled={true}
               className={`mt-4 px-4 py-2 bg-white ${color[1]} font-semibold rounded hover:bg-blue-100 transition`}
               onClick={handleButton}
           >
@@ -68,25 +76,7 @@ export default function TaskView({ task }) {
           </svg>
           </div>
           <div className="absolute bottom-3 right-7 scale-125 rotate-[-20deg]">
-            <svg width="150" height="180" viewBox="0 0 621 789" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <rect x="88.216" y="136.289" width="443.753" height="217.011" rx="16.0393" stroke="white" strokeWidth="32.0785"/>
-              <rect x="16.0393" y="411.147" width="588.107" height="360.882" rx="16.0393" stroke="white" strokeWidth="32.0785"/>
-              <line x1="80.1963" y1="288.881" x2="537.315" y2="288.881" stroke="white" strokeWidth="32.0785"/>
-              <line x1="168.412" y1="279.152" x2="168.412" y2="369.34" stroke="white" strokeWidth="32.0785"/>
-              <line x1="518.604" y1="505.63" x2="181.779" y2="505.63" stroke="white" strokeWidth="32.0785"/>
-              <line x1="518.603" y1="589.375" x2="181.778" y2="589.375" stroke="white" strokeWidth="32.0785"/>
-              <line x1="518.604" y1="670.974" x2="181.779" y2="670.974" stroke="white" strokeWidth="32.0785"/>
-              <line x1="459.793" y1="279.152" x2="459.793" y2="369.34" stroke="white" strokeWidth="32.0785"/>
-              <line x1="358.211" y1="279.152" x2="358.211" y2="369.34" stroke="white" strokeWidth="32.0785"/>
-              <line x1="261.975" y1="279.152" x2="261.975" y2="369.34" stroke="white" strokeWidth="32.0785"/>
-              <line x1="310.093" y1="47.241" x2="310.093" y2="137.429" stroke="white" strokeWidth="32.0785"/>
-              <ellipse cx="169.749" cy="209.364" rx="41.4348" ry="33.2836" fill="white"/>
-              <ellipse cx="114.949" cy="502.474" rx="29.4053" ry="23.6206" fill="white"/>
-              <ellipse cx="114.949" cy="586.22" rx="29.4053" ry="23.6206" fill="white"/>
-              <ellipse cx="114.948" cy="667.818" rx="29.4053" ry="23.6206" fill="white"/>
-              <ellipse cx="447.763" cy="209.364" rx="41.4348" ry="33.2836" fill="white"/>
-              <ellipse cx="310.093" cy="33.2836" rx="42.7714" ry="33.2836" fill="white"/>
-            </svg>
+            <ListBuddyTaskSVG />
           </div>
       </div>
   );
